+++ src/main/java/com/group2/smart_cafe_backend/models/Tables.java
@@ -10,7 +10,7 @@ public class Tables {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int tableId;
+    private Long tableId;
 
     @Column(nullable = false, unique = true, length = 50)
     private String code;
